requirements.txt
ollama==0.5.1
datasets==3.6.0
transformers==4.52.4
sentence-transformers==4.1.0
torch==2.7.0
fastapi==0.115.12
uvicorn==0.34.3
gradio==5.32.1
faiss-cpu==1.11.0
pydantic
pandas
numpy